update dependencies and tests
